+++ docs/HISTORY.rst
@@ -1,6 +1,14 @@
 Changelog
 =========
 
+0.2.0 (Unreleased)
+------------------
+
+* Switched to new embed mode (using ``iframe``)
+  [keul]
+* Video provided in https
+  [keul]
+
 0.1.0 (2011-08-31)
 ------------------